// easyprufung-dashboard-app/src/store/models/user/project/projectIconConfiguration.interface.ts
export interface IProjectIconConfiguration {
    content: string;
    size: number;
    rotation: number;
    strokeWidth: number;
    strokeColor: string;
    fillColor: string;
    fillOpacity: number;
    backgroundColor: string;
    backgroundRounded: number;
    backgroundPadding: number;
    isGradientBackGround: boolean;
}